+++ basics/processing-instructions/steel/api/src/lib.rs
@@ -1,0 +1,9 @@
+mod instruction;
+
+use steel::*;
+
+declare_id!("z7msBPQHDJjTvdQRoEcKyENgXDhSRYeHieN1ZMTqo35");
+
+pub mod prelude {
+    pub use crate::instruction::*;
+}
