+++ basics/processing-instructions/steel/program/Cargo.toml
@@ -4,10 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-solana-program = "1.18.17"
-steel = "2.1"
-bytemuck = "1.4"
-num_enum = "0.7"
+steel.workspace = true
+processing-instructions-steel-api = {path = "../api"}
 
 [lib]
 crate-type = ["cdylib", "lib"]
